Prototyped: - Death mechanic - Shadow mechanic (unconscious) - Boss visuals
Fixed
- Outline stuck (still not perfect tho)
- Ai (adjusted)

Added
- More info-text events (shield, extra atks)
- Initial Quests
- 4 heroes
- 1 final level

+++ Dueling-Club/src/main/java/eidolons/game/battlecraft/logic/battlefield/DroppedItemManager.java
@@ -2,9 +2,11 @@ package eidolons.game.battlecraft.logic.battlefield;
 
 import eidolons.content.PROPS;
 import eidolons.entity.item.DC_HeroItemObj;
+import eidolons.entity.item.DC_HeroSlotItem;
 import eidolons.entity.obj.unit.Unit;
 import eidolons.game.core.game.DC_Game;
 import eidolons.libgdx.gui.panels.dc.inventory.InventoryClickHandler.CONTAINER;
+import eidolons.swing.renderers.SlotItem;
 import eidolons.system.ObjUtilities;
 import main.content.enums.entity.ItemEnums;
 import main.entity.Entity;
@@ -21,7 +23,7 @@ import java.util.List;
 
 public class DroppedItemManager {
 
-    private static final List<DC_HeroItemObj> VOID =     new ArrayList<>();
+    private static final List<DC_HeroItemObj> VOID = new ArrayList<>();
     List<DC_HeroItemObj>[][] itemMap;
     private DC_Game game;
 
@@ -31,8 +33,8 @@ public class DroppedItemManager {
 
     public void init() {
         itemMap = new List
-         [game.getDungeon().getCellsX()]
-         [game.getDungeon().getCellsY()];
+                [game.getDungeon().getCellsX()]
+                [game.getDungeon().getCellsY()];
 
     }
 
@@ -42,28 +44,38 @@ public class DroppedItemManager {
     }
 
     public void dropDead(Unit unit) {
-        unit.unequip(ItemEnums.ITEM_SLOT.ARMOR);
-        unit.unequip(ItemEnums.ITEM_SLOT.MAIN_HAND);
-        unit.unequip(ItemEnums.ITEM_SLOT.OFF_HAND);
-        // drop natural weapons?
-
+        if (!unit.isRevenant()) {
+            unit.unequip(ItemEnums.ITEM_SLOT.ARMOR);
+            unit.unequip(ItemEnums.ITEM_SLOT.MAIN_HAND);
+            unit.unequip(ItemEnums.ITEM_SLOT.OFF_HAND);
+            // drop natural weapons?
+        }
         for (DC_HeroItemObj item : unit.getInventory()) {
             if (checkLootDrops(item, unit))
                 drop(item, unit.getCoordinates());
-            else{
-                destroyItem(item);
+            else {
+                if (!unit.isMine())
+                    destroyItem(item);
             }
         }
-        for (DC_HeroItemObj item : unit.getQuickItems()) {
-            drop(item, unit.getCoordinates());
+        if (!unit.isRevenant()) {
+            for (DC_HeroItemObj item : unit.getQuickItems()) {
+                drop(item, unit.getCoordinates());
+            }
+            for (DC_HeroItemObj item : unit.getJewelry()) {
+                drop(item, unit.getCoordinates());
+            }
         }
-        for (DC_HeroItemObj item : unit.getJewelry()) {
-            drop(item, unit.getCoordinates());
-        }
-
     }
 
     private boolean checkLootDrops(DC_HeroItemObj item, Unit unit) {
+        if (unit.isMine()) {
+            //check if not original!
+            if (item instanceof DC_HeroSlotItem)
+                if (item.getOriginalUnit() == unit)
+                    return false;
+            return true;
+        }
         return
                 RandomWizard.chance(unit.getGame().getMetaMaster().getLootMaster().
                         getChanceForOwnedItemToDrop(unit, item));
@@ -93,11 +105,13 @@ public class DroppedItemManager {
 
     public void reset(int i, int j) {
         itemMap[i][j] =
-         getItems(Coordinates.get(i, j));
+                getItems(Coordinates.get(i, j));
     }
+
     public List<DC_HeroItemObj> getDroppedItems(Obj cell) {
         return getDroppedItems(cell.getCoordinates());
     }
+
     public List<DC_HeroItemObj> getDroppedItems(Coordinates coordinates) {
         return itemMap[coordinates.getX()][coordinates.getY()];
     }
@@ -109,7 +123,7 @@ public class DroppedItemManager {
 
     private List<DC_HeroItemObj> getItems(Obj cell) {
         if (cell == null) {
-            return     VOID ;
+            return VOID;
         }
         List<DC_HeroItemObj> list = new ArrayList<>();
         for (String id : ContainerUtils.open(cell.getProperty(PROPS.DROPPED_ITEMS))) {
@@ -124,17 +138,17 @@ public class DroppedItemManager {
 
     public void remove(DC_HeroItemObj item, Unit heroObj) {
         game.getCellByCoordinate(heroObj.getCoordinates()).removeProperty(PROPS.DROPPED_ITEMS,
-         "" + item.getId());
+                "" + item.getId());
     }
 
     public void remove(Unit heroObj, Entity item) {
         game.getCellByCoordinate(heroObj.getCoordinates()).removeProperty(PROPS.DROPPED_ITEMS,
-         "" + item.getId());
+                "" + item.getId());
     }
 
     public void drop(DC_HeroItemObj item, Coordinates c) {
         game.getCellByCoordinate(c).addProperty(PROPS.DROPPED_ITEMS,
-         "" + item.getId());
+                "" + item.getId());
         item.setCoordinates(c);
         item.setContainer(CONTAINER.UNASSIGNED);
     }
@@ -143,6 +157,7 @@ public class DroppedItemManager {
         cell.removeProperty(PROPS.DROPPED_ITEMS, "" + item.getId());
         return true;
     }
+
     public void pickedUp(Obj item) {
         if (!(item instanceof DC_HeroItemObj)) {
             return;
@@ -164,7 +179,7 @@ public class DroppedItemManager {
     public boolean pickUp(Unit hero, Entity type) {
         Obj cell = game.getCellByCoordinate(hero.getCoordinates());
         DC_HeroItemObj item = (DC_HeroItemObj) ObjUtilities.findObjByType(type,
-         getDroppedItems(cell));
+                getDroppedItems(cell));
         if (item == null) {
             return false;
         }
@@ -175,7 +190,7 @@ public class DroppedItemManager {
     public DC_HeroItemObj findDroppedItem(String typeName, Coordinates coordinates) {
         Obj cell = game.getCellByCoordinate(coordinates);
         Obj item = new ListMaster<Obj>()
-         .findType(typeName, new ArrayList<>(getDroppedItems(cell)));
+                .findType(typeName, new ArrayList<>(getDroppedItems(cell)));
         return (DC_HeroItemObj) item;
     }
